Backend/requirements.txt
asgiref==3.7.2
Django==4.2.6
django-cors-headers==4.3.0
djangorestframework==3.14.0
joblib==1.3.2
numpy==1.26.1
pytz==2023.3.post1
scikit-learn==0.24.1
scipy==1.11.3
sklearn==0.0.post5
sqlparse==0.4.4
threadpoolctl==3.2.0
typing_extensions==4.8.0
tzdata==2023.3
